+++ src/app/engines/markdown.engine.ts
@@ -13,7 +13,30 @@ export class MarkdownEngine {
             return `<pre><code class="hljs ${language}">${highlighted}</code></pre>`;
         };
 
-        marked.setOptions({ renderer });
+        renderer.table = (header, body) => {
+            return '<table class="table table-bordered compodoc-table">\n'
+                + '<thead>\n'
+                + header
+                + '</thead>\n'
+                + '<tbody>\n'
+                + body
+                + '</tbody>\n'
+                + '</table>\n';
+        }
+
+        renderer.image = function(href, title, text) {
+          var out = '<img src="' + href + '" alt="' + text + '" class="img-responsive"';
+          if (title) {
+            out += ' title="' + title + '"';
+          }
+          out += this.options.xhtml ? '/>' : '>';
+          return out;
+        };
+
+        marked.setOptions({
+            renderer: renderer,
+            breaks: true
+        });
     }
     getReadmeFile() {
         return new Promise(function(resolve, reject) {