+++ src/utils/defaults.ts
@@ -1,12 +1,13 @@
 export const COMPODOC_DEFAULTS = {
+    inputDir: '',
     title: 'Application documentation',
     additionalEntryName: 'Additional documentation',
     additionalEntryPath: 'additional-documentation',
-    inputDir: '',
     folder: './documentation/',
     port: 8080,
     theme: 'gitbook',
     base: '/',
     disableSourceCode: false,
-    disableGraph: false
+    disableGraph: false,
+    disableCoverage: false
 }